Funciona el numero de votos

+++ src/main/java/models/daos/VotoDao.java
@@ -6,7 +6,7 @@ public interface VotoDao extends GenericDao<Voto,Integer> {
 
 	public void BorrarVotosdeTema(int idTema);
 
-	public int getNumVotos(int idTema);
+	public long getNumVotos(int idTema);
 
 	public int getValorMedia(int idTema);
